move V8 file into V8 subdirectory
