+++ arangod/V8Server/V8DealerFeature.cpp
@@ -34,8 +34,6 @@
 #include "Agency/v8-agency.h"
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "ApplicationFeatures/HttpEndpointProvider.h"
-#include "ApplicationFeatures/V8PlatformFeature.h"
-#include "V8/V8SecurityFeature.h"
 #include "Basics/ArangoGlobalContext.h"
 #include "Basics/FileUtils.h"
 #include "Basics/ScopeGuard.h"
@@ -67,6 +65,8 @@
 #include "Transaction/V8Context.h"
 #include "Utilities/NameValidator.h"
 #include "V8/JavaScriptSecurityContext.h"
+#include "V8/V8PlatformFeature.h"
+#include "V8/V8SecurityFeature.h"
 #include "V8/v8-buffer.h"
 #include "V8/v8-conv.h"
 #include "V8/v8-globals.h"
