+++ lib/V8/v8-globals.h
@@ -37,11 +37,11 @@
 
 #include <v8.h>
 
-#include "ApplicationFeatures/V8PlatformFeature.h"
 #include "Basics/Common.h"
 #include "Basics/StringBuffer.h"
 #include "Basics/operating-system.h"
 #include "V8/JavaScriptSecurityContext.h"
+#include "V8/V8PlatformFeature.h"
 
 struct TRI_v8_global_t;
 struct TRI_vocbase_t;
