+++ lib/V8/v8-vpack.cpp
@@ -31,10 +31,10 @@
 #include <velocypack/Iterator.h>
 #include <velocypack/Slice.h>
 
-#include "ApplicationFeatures/V8PlatformFeature.h"
 #include "Basics/Exceptions.h"
 #include "Basics/VelocyPackHelper.h"
 #include "Basics/debugging.h"
+#include "V8/V8PlatformFeature.h"
 #include "V8/v8-conv.h"
 #include "V8/v8-utils.h"
